separate tns retrieval and submission queues (#4392)
This PR creates separate TNS submission and retrieval queues.
