+++ config.yaml.defaults
@@ -652,8 +652,9 @@ ports:
   facility_queue: 64510
   notification_queue: 64610
   observation_plan_queue: 64710
-  tns_queue: 64810
+  tns_retrieval_queue: 64810
   thumbnail_queue: 64910
+  tns_submission_queue: 65010
 
 gcn:
   server: gcn.nasa.gov
