+++ services/tns_retrieval_queue/supervisor.conf
@@ -1,0 +1,5 @@
+[program:tns_retrieval_queue]
+command=/usr/bin/env python services/tns_retrieval_queue/tns_retrieval_queue.py %(ENV_FLAGS)s
+environment=PYTHONPATH=".",PYTHONUNBUFFERED="1"
+stdout_logfile=log/tns_retrieval.log
+redirect_stderr=true
