+++ services/tns_submission_queue/supervisor.conf
@@ -1,0 +1,5 @@
+[program:tns_submission_queue]
+command=/usr/bin/env python services/tns_submission_queue/tns_submission_queue.py %(ENV_FLAGS)s
+environment=PYTHONPATH=".",PYTHONUNBUFFERED="1"
+stdout_logfile=log/tns_submission.log
+redirect_stderr=true
