+++ skyportal/handlers/api/tns.py
@@ -1,13 +1,11 @@
 import asyncio
 import json
 import tempfile
-import time
 import urllib
 
 import arrow
 import astropy.units as u
 import requests
-import sqlalchemy as sa
 from astropy.time import Time, TimeDelta
 from marshmallow.exceptions import ValidationError
 from sqlalchemy.orm import scoped_session, sessionmaker
@@ -15,31 +13,23 @@ from tornado.ioloop import IOLoop
 
 from baselayer.app.access import auth_or_token, permissions
 from baselayer.app.env import load_env
-from baselayer.app.flow import Flow
 from baselayer.app.model_util import recursive_to_dict
 from baselayer.log import make_log
 
 from ...models import (
-    DBSession,
     Group,
     Obj,
     Spectrum,
     SpectrumObserver,
     SpectrumReducer,
     TNSRobot,
-    User,
 )
 from ...utils.tns import (
     get_IAUname,
-    get_recent_TNS,
+    get_tns,
     post_tns,
-    read_tns_photometry,
-    read_tns_spectrum,
 )
 from ..base import BaseHandler
-from .photometry import add_external_photometry
-from .source import post_source
-from .spectrum import post_spectrum
 
 _, cfg = load_env()
 
@@ -48,8 +38,6 @@ Session = scoped_session(sessionmaker())
 TNS_URL = cfg['app.tns.endpoint']
 upload_url = urllib.parse.urljoin(TNS_URL, 'api/file-upload')
 report_url = urllib.parse.urljoin(TNS_URL, 'api/bulk-report')
-search_url = urllib.parse.urljoin(TNS_URL, 'api/get/search')
-object_url = urllib.parse.urljoin(TNS_URL, 'api/get/object')
 
 log = make_log('api/tns')
 
@@ -328,258 +316,6 @@ class TNSRobotHandler(BaseHandler):
             return self.success()
 
 
-def tns_bulk_retrieval(
-    start_date,
-    tnsrobot_id,
-    user_id,
-    group_ids=None,
-    include_photometry=False,
-    include_spectra=False,
-):
-
-    """Retrieve objects from TNS.
-    start_date : str
-        ISO-based start time
-    tnsrobot_id : int
-        TNSRobot ID
-    user_id : int
-        SkyPortal ID of User retrieving from TNS
-    group_ids : List[int]
-        List of groups to post TNS sources to
-    include_photometry: boolean
-        Include photometry available on TNS
-    include_spectra : boolean
-        Include spectra available on TNS
-    """
-
-    if Session.registry.has():
-        session = Session()
-    else:
-        session = Session(bind=DBSession.session_factory.kw["bind"])
-
-    user = session.scalar(sa.select(User).where(User.id == user_id))
-    if group_ids is None:
-        group_ids = [g.id for g in user.accessible_groups]
-
-    try:
-        tnsrobot = session.scalars(
-            TNSRobot.select(user).where(TNSRobot.id == tnsrobot_id)
-        ).first()
-        if tnsrobot is None:
-            raise ValueError(f'No TNSRobot available with ID {tnsrobot_id}')
-
-        altdata = tnsrobot.altdata
-        if not altdata:
-            raise ValueError('Missing TNS information.')
-        if 'api_key' not in altdata:
-            raise ValueError('Missing TNS API key.')
-
-        tns_headers = {
-            'User-Agent': f'tns_marker{{"tns_id":{tnsrobot.bot_id},"type":"bot", "name":"{tnsrobot.bot_name}"}}'
-        }
-
-        tns_sources = get_recent_TNS(altdata['api_key'], tns_headers, start_date)
-        if len(tns_sources) == 0:
-            raise ValueError(f'No objects posted to TNS since {start_date}.')
-
-        for source in tns_sources:
-            s = session.scalars(Obj.select(user).where(Obj.id == source['id'])).first()
-            if s is None:
-                log(f"Posting {source['id']} as source")
-                source['group_ids'] = group_ids
-                post_source(source, user_id, session)
-
-            tns_retrieval(
-                source['id'],
-                tnsrobot_id,
-                user_id,
-                include_photometry=include_photometry,
-                include_spectra=include_spectra,
-                parent_session=session,
-            )
-        session.commit()
-
-    except Exception as e:
-        log(f"Unable to retrieve TNS report for objects since {start_date}: {e}")
-    finally:
-        session.close()
-        Session.remove()
-
-
-def tns_retrieval(
-    obj_id,
-    tnsrobot_id,
-    user_id,
-    include_photometry=False,
-    include_spectra=False,
-    parent_session=None,
-):
-    """Retrieve object from TNS.
-    obj_id : str
-        Object ID
-    tnsrobot_id : int
-        TNSRobot ID
-    user_id : int
-        SkyPortal ID of User retrieving from TNS
-    include_photometry: boolean
-        Include photometry available on TNS
-    include_spectra : boolean
-        Include spectra available on TNS
-    """
-
-    if parent_session is None:
-        if Session.registry.has():
-            session = Session()
-        else:
-            session = Session(bind=DBSession.session_factory.kw["bind"])
-    else:
-        session = parent_session
-
-    flow = Flow()
-
-    user = session.scalar(sa.select(User).where(User.id == user_id))
-
-    try:
-        obj = session.scalars(Obj.select(user).where(Obj.id == obj_id)).first()
-        if obj is None:
-            raise ValueError(f'No object available with ID {obj_id}')
-
-        tnsrobot = session.scalars(
-            TNSRobot.select(user).where(TNSRobot.id == tnsrobot_id)
-        ).first()
-        if tnsrobot is None:
-            raise ValueError(f'No TNSRobot available with ID {tnsrobot_id}')
-
-        altdata = tnsrobot.altdata
-        if not altdata:
-            raise ValueError('Missing TNS information.')
-        if 'api_key' not in altdata:
-            raise ValueError('Missing TNS API key.')
-
-        tns_headers = {
-            'User-Agent': f'tns_marker{{"tns_id":{tnsrobot.bot_id},"type":"bot", "name":"{tnsrobot.bot_name}"}}'
-        }
-
-        _, tns_name = get_IAUname(
-            altdata['api_key'], tns_headers, ra=obj.ra, dec=obj.dec
-        )
-        if tns_name is None:
-            raise ValueError(f'{obj_id} not yet posted to TNS.')
-
-        obj.tns_name = tns_name
-
-        data = {
-            'api_key': altdata['api_key'],
-            'data': json.dumps(
-                {
-                    "objname": tns_name,
-                    "photometry": "1" if include_photometry else "0",
-                    "spectra": "1" if include_spectra else "0",
-                }
-            ),
-        }
-
-        r = requests.post(
-            object_url,
-            headers=tns_headers,
-            data=data,
-            allow_redirects=True,
-            stream=True,
-            timeout=10,
-        )
-
-        count = 0
-        count_limit = 5
-        while r.status_code == 429 and count < count_limit:
-            log(
-                f'TNS request rate limited: {str(r.json())}.  Waiting 30 seconds to try again.'
-            )
-            time.sleep(30)
-            r = requests.post(
-                object_url,
-                headers=tns_headers,
-                data=data,
-                allow_redirects=True,
-                stream=True,
-                timeout=10,
-            )
-            count += 1
-
-        if count == count_limit:
-            raise ValueError('TNS request failed: request rate exceeded.')
-
-        if r.status_code == 200:
-            source_data = r.json().get("data", dict()).get("reply", dict())
-            if source_data:
-                obj.tns_info = source_data
-                group_ids = [g.id for g in user.accessible_groups]
-
-                if include_photometry and 'photometry' in source_data:
-                    photometry = source_data['photometry']
-
-                    failed_photometry = []
-                    failed_photometry_errors = []
-
-                    for phot in photometry:
-                        try:
-                            df, instrument_id = read_tns_photometry(phot, session)
-                            data_out = {
-                                'obj_id': obj_id,
-                                'instrument_id': instrument_id,
-                                'group_ids': group_ids,
-                                **df.to_dict(orient='list'),
-                            }
-                            add_external_photometry(data_out, user)
-                        except Exception as e:
-                            failed_photometry.append(phot)
-                            failed_photometry_errors.append(str(e))
-                            log(f'Cannot read TNS photometry {str(phot)}: {str(e)}')
-                            continue
-                    if len(failed_photometry) > 0:
-                        log(
-                            f'Failed to retrieve {len(failed_photometry)}/{len(photometry)} TNS photometry for {obj_id} from TNS as {tns_name}: {str(list(set(failed_photometry_errors)))}'
-                        )
-                if include_spectra and 'spectra' in source_data:
-                    group_ids = [g.id for g in user.accessible_groups]
-                    spectra = source_data['spectra']
-
-                    failed_spectra = []
-                    failed_spectra_errors = []
-
-                    for spectrum in spectra:
-                        try:
-                            data = read_tns_spectrum(spectrum, session)
-                        except Exception as e:
-                            log(f'Cannot read TNS spectrum {str(spectrum)}: {str(e)}')
-                            continue
-                        data["obj_id"] = obj_id
-                        data["group_ids"] = group_ids
-                        post_spectrum(data, user_id, session)
-
-                    if len(failed_spectra) > 0:
-                        log(
-                            f'Failed to retrieve {len(failed_spectra)}/{len(spectra)} TNS spectra for {obj_id} from TNS as {tns_name}: {str(list(set(failed_spectra_errors)))}'
-                        )
-
-            log(f'Successfully retrieved {obj_id} from TNS as {tns_name}')
-        else:
-            log(f'Failed to retrieve {obj_id} from TNS: {r.content}')
-        session.commit()
-
-        flow.push(
-            '*',
-            'skyportal/REFRESH_SOURCE',
-            payload={'obj_key': obj.internal_key},
-        )
-
-    except Exception as e:
-        log(f"Unable to retrieve TNS report for {obj_id}: {e}")
-    finally:
-        if parent_session is not None:
-            session.close()
-            Session.remove()
-
-
 class BulkTNSHandler(BaseHandler):
     @auth_or_token
     def post(self):
@@ -664,15 +400,16 @@ class BulkTNSHandler(BaseHandler):
             except Exception:
                 loop = asyncio.new_event_loop()
                 asyncio.set_event_loop(loop)
+
             IOLoop.current().run_in_executor(
                 None,
-                lambda: tns_bulk_retrieval(
-                    start_date.isot,
+                lambda: get_tns(
                     tnsrobot.id,
                     self.associated_user_object.id,
-                    group_ids=group_ids,
                     include_photometry=include_photometry,
                     include_spectra=include_spectra,
+                    start_date=start_date.isot,
+                    group_ids=group_ids,
                 ),
             )
 
@@ -735,14 +472,15 @@ class ObjTNSHandler(BaseHandler):
             except Exception:
                 loop = asyncio.new_event_loop()
                 asyncio.set_event_loop(loop)
+
             IOLoop.current().run_in_executor(
                 None,
-                lambda: tns_retrieval(
-                    obj.id,
+                lambda: get_tns(
                     tnsrobot.id,
                     self.associated_user_object.id,
                     include_photometry=include_photometry,
                     include_spectra=include_spectra,
+                    obj_id=obj.id,
                 ),
             )
